fix index.ts
- refactor #155
